added Sidebar component
